extract recent used custom date logic to be an independent component
